Store loading informations done
