+++ src/config/config_definitions.hpp
@@ -50,7 +50,6 @@ enum booleanConfig_t {
 	FORCE_MONSTERTYPE_LOAD,
 	HOUSE_OWNED_BY_ACCOUNT,
 	CLEAN_PROTECTION_ZONES,
-	STOREMODULES,
 	ALLOW_BLOCK_SPAWN,
 	ONLY_INVITED_CAN_MOVE_HOUSE_ITEMS,
 	WEATHER_RAIN,
