+++ src/lua/functions/core/game/config_functions.hpp
@@ -58,7 +58,6 @@ class ConfigFunctions final : LuaScriptInterface {
 			registerEnumIn(L, "configKeys", REMOVE_WEAPON_AMMO)
 			registerEnumIn(L, "configKeys", REMOVE_WEAPON_CHARGES)
 			registerEnumIn(L, "configKeys", REMOVE_POTION_CHARGES)
-			registerEnumIn(L, "configKeys", STOREMODULES)
 			registerEnumIn(L, "configKeys", WEATHER_RAIN)
 			registerEnumIn(L, "configKeys", WEATHER_THUNDER)
 			registerEnumIn(L, "configKeys", FREE_QUESTS)
